listener added to menu and contact buttons
