+++ src/assets/style.css
@@ -4,8 +4,8 @@ h1 {
 
 .menuImg {
     margin: 0 auto;
-    /* width: 600px; */
-    /* height: 600px; */
+    width: 800px;
+    height: 600px; 
 }
 
 .contactDiv {
