+++ src/modules/contactButton.js
@@ -1,0 +1,5 @@
+
+contactButton.addEventListener('click', e => {
+    content.lastElementChild.remove()
+    contactCreation()
+})
