+++ src/modules/menuButton.js
@@ -1,0 +1,7 @@
+
+
+
+menuButton.addEventListener('click', e => {
+    content.lastElementChild.remove()
+    menuCreation()
+})
